+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/exchangescanner.py
@@ -29,7 +29,6 @@ from os2datascanner.engine2.model.core import SourceManager
 
 from os2datascanner.projects.grants.models import GraphGrant, EWSGrant
 from ....organizations.models.account import Account
-from ....organizations.models.aliases import AliasType
 from ...utils import upload_path_exchange_users
 from .scanner import Scanner
 from os2datascanner.engine2.rules.dict_lookup import EmailHeaderRule
@@ -136,10 +135,8 @@ class ExchangeScanner(Scanner):
             for account in covered_accounts:
                 # Only try to scan mail addresses that belong to the domain
                 # associated with this scanner
-                for alias in account.aliases.filter(
-                        _alias_type=AliasType.EMAIL,
-                        _value__iendswith=self.mail_domain):
-                    user_mail_address: str = alias.value
+                user_mail_address: str = account.email
+                if user_mail_address:
                     local_part = user_mail_address.split("@", maxsplit=1)[0]
                     yield (account, _make_source(user=local_part))
         elif self.userlist: